+++ firebase_apis/test_api/src/main/java/com/google/api/services/testing/model/AndroidModel.java
@@ -134,7 +134,8 @@ public final class AndroidModel extends com.google.api.client.json.GenericJson {
   private java.util.List<java.lang.String> tags;
 
   /**
-   * URL of a thumbnail image of the device.
+   * URL of a thumbnail image (photo) of the device. e.g.
+   * https://lh3.googleusercontent.com/90WcauuJiCYABEl8U0lcZeuS5STUbf2yW...
    * The value may be {@code null}.
    */
   @com.google.api.client.util.Key
@@ -391,7 +392,8 @@ public final class AndroidModel extends com.google.api.client.json.GenericJson {
   }
 
   /**
-   * URL of a thumbnail image of the device.
+   * URL of a thumbnail image (photo) of the device. e.g.
+   * https://lh3.googleusercontent.com/90WcauuJiCYABEl8U0lcZeuS5STUbf2yW...
    * @return value or {@code null} for none
    */
   public java.lang.String getThumbnailUrl() {
@@ -399,7 +401,8 @@ public final class AndroidModel extends com.google.api.client.json.GenericJson {
   }
 
   /**
-   * URL of a thumbnail image of the device.
+   * URL of a thumbnail image (photo) of the device. e.g.
+   * https://lh3.googleusercontent.com/90WcauuJiCYABEl8U0lcZeuS5STUbf2yW...
    * @param thumbnailUrl thumbnailUrl or {@code null} for none
    */
   public AndroidModel setThumbnailUrl(java.lang.String thumbnailUrl) {